refactor: 추상화에 따른 클로바 로직 분리
#30
