+++ src/main/java/bookhive/bookhiveserver/domain/ai/client/clova/ClovaSearchApiClient.java
@@ -2,7 +2,7 @@ package bookhive.bookhiveserver.domain.ai.client.clova;
 
 import bookhive.bookhiveserver.domain.ai.dto.request.clova.ClovaMessage;
 import bookhive.bookhiveserver.domain.ai.dto.request.clova.ClovaRequest;
-import bookhive.bookhiveserver.domain.ai.dto.response.KeywordsResponse;
+import bookhive.bookhiveserver.domain.ai.dto.response.clova.ClovaKeywordsResponse;
 import bookhive.bookhiveserver.domain.ai.dto.response.clova.ClovaResponse;
 import bookhive.bookhiveserver.domain.ai.dto.response.clova.ClovaSearchTypeResponse;
 import bookhive.bookhiveserver.global.exception.ErrorMessage;
@@ -94,7 +94,7 @@ public class ClovaSearchApiClient {
         }
 }
 
-    public KeywordsResponse extractKeywords(String question, String originTags) {
+    public ClovaKeywordsResponse extractKeywords(String question, String originTags) {
 
         List<ClovaMessage> messages = List.of(
                 new ClovaMessage("system",
@@ -155,7 +155,7 @@ public class ClovaSearchApiClient {
 //        System.out.println("AI 검색 결과: \n" + jsonString);
 
         try {
-            return objectMapper.readValue(jsonString, KeywordsResponse.class);
+            return objectMapper.readValue(jsonString, ClovaKeywordsResponse.class);
         } catch (JsonProcessingException e) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, ErrorMessage.JSON_PARSE_ERROR.toString(), e);
         }
