+++ src/main/java/bookhive/bookhiveserver/domain/ai/dto/response/clova/ClovaKeywordsResponse.java
@@ -1,4 +1,4 @@
-package bookhive.bookhiveserver.domain.ai.dto.response;
+package bookhive.bookhiveserver.domain.ai.dto.response.clova;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
@@ -6,6 +6,6 @@ import lombok.Getter;
 
 @Getter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class KeywordsResponse {
+public class ClovaKeywordsResponse {
     private List<String> keywords;
 }
